Relocate derive_photon_feedback_outcome + injection message (Phase 8 batch 3, #688)
Issue #688 (parent #680, Phase 8) batch 3: migrate the photon feedback
outcome derivation core + injection message builder from `turn.rs` to
`photon_feedback_derive.rs`.

Migrated as free fns in `photon_feedback_derive.rs`:

* `pub(crate) fn derive_photon_feedback_outcome(&PhotonOutcomeInputs)
    -> PhotonFeedbackOutcome` (the 7-case classifier — Case A shadow
    short-circuit, Case B zero-adoption short-circuit, Case C
    safety_violation, Case D failure, Case E success, Case F
    no_progress_despite_inject, Case G fallback)
* `fn is_eligible_failure_kind(&FeedbackKind) -> bool` (module-private,
    Case D allowlist — 9 variants from `is_eligible_for_reminder`
    minus `UnsafeCommandBlocked`)
* `pub fn build_photon_injection_message(response, shadow_mode)
    -> Option<ConversationMessage>` (re-exported via `loop_run.rs`
    for `tests/photon_turn_hook_smoke.rs`)
* `#[cfg(test)] impl<'a> PhotonOutcomeInputs<'a> { test_default() }`
    (no-signal fixture builder, `pub(crate)`)

`photon_feedback_derive.rs` imports gain
`crate::session::store::ConversationMessage`.

`turn.rs` adjustments:

* Removed the 3 fn definitions + `#[cfg(test)]` impl block
  (originally at lines 585-741 + 763-781).
* Updated the existing `use super::photon_feedback_derive::{...}`
  import to add `PhotonFeedbackOutcome` + `derive_photon_feedback_outcome`
  (the new caller-facing imports) and drop
  `PHOTON_OUTCOME_DETAIL_NO_PROGRESS_DESPITE_INJECT` + `case_f_condition_met`
  (only the test mod needs them now and it imports via the dedicated
  `super::super::photon_feedback_derive::*` path).
* Test mod `derive_photon_feedback_outcome_tests` rewrote its imports
  to `super::super::photon_feedback_derive::*`.

`loop_run.rs` adjustments:

* Split the existing `pub use turn::{
    MAX_PHOTON_CONTEXT_PACK_PROMPT_BYTES, build_photon_injection_message,
    truncate_photon_context_pack
  };` so `build_photon_injection_message` is now re-exported from
  `photon_feedback_derive` while the other two stay on `turn`.
  `tests/photon_turn_hook_smoke.rs` import path is unchanged
  (`anvil::agent::loop_run::build_photon_injection_message`).

LOC delta:

* `turn.rs`: 29,641 → 29,439 (-202 LOC)
* `photon_feedback_derive.rs`: 310 → 511 (+201 LOC)
* Cumulative `turn.rs`: 39,489 → 29,439 (-10,050 LOC, -25.5%)

This batch crosses the **−10K LOC / −25%** milestone for the meta-issue
#680 reduction target (final goal: ≤10,000 LOC).

`pub(crate)` for the outcome derivation core, `pub` for the
injection-message builder (matching pre-extraction visibility). `turn.rs`
remains the only in-crate consumer of the types.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
